# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Education Mini Program API
  version: 0.1.0
  description: >-
    Shared API contract for the first-phase education mini program. Successful
    responses use a data envelope and failures preserve their HTTP status.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Environment-relative service root
...
security:
  - bearerAuth: []
paths:
...
  /teachers/me/lesson-sessions/{lessonSessionId}/attendance:
    put:
      operationId: saveTeacherAttendance
      summary: Save a complete attendance draft for an assigned lesson
      tags:
        - Teacher
      parameters:
        - $ref: '#/components/parameters/LessonSessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceDraft'
      responses:
        '200':
          description: Attendance draft saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherLessonSessionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /teachers/me/lesson-sessions/{lessonSessionId}/complete:
    post:
      operationId: completeTeacherLessonSession
      summary: Complete an assigned lesson and consume student lesson units
      tags:
        - Teacher
      parameters:
        - $ref: '#/components/parameters/LessonSessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteLessonRequest'
      responses:
        '200':
          description: Lesson completion result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteLessonEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /teachers/me/lesson-sessions/{lessonSessionId}/reverse:
    post:
      operationId: reverseTeacherLessonSession
      summary: Reverse a completed lesson within the teacher reversal window
      tags:
        - Teacher
      parameters:
        - $ref: '#/components/parameters/LessonSessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReverseLessonRequest'
      responses:
        '200':
          description: Lesson reversal result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteLessonEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the authentication module.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Caller-generated key used by mutation endpoints to make retries safe.
        Keys contain 8-128 letters, numbers, dots, underscores, colons, or
        hyphens.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]{8,128}$'
      example: lesson-confirm-20260827-0001
...
    LessonSessionId:
      name: lessonSessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      description: Correlation identifier matching the response body requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
...
  responses:
...
    BadRequest:
      description: Request validation failed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is required or the session is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The authenticated user lacks the required role or data scope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The requested state transition conflicts with current state.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ResourceNotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missingResource:
              value:
                code: RESOURCE_NOT_FOUND
                message: Cannot GET /resource
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                code: INTERNAL_SERVER_ERROR
                message: Internal server error
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Common wrapper for every successful response.
      required:
        - data
        - requestId
      properties:
        data:
          description: Endpoint-specific response payload.
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
...
    ErrorEnvelope:
      type: object
      description: Common response body for failures; the HTTP status is preserved.
      required:
        - code
        - message
        - requestId
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
    ErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - VALIDATION_FAILED
        - INVALID_IDEMPOTENCY_KEY
        - UNAUTHORIZED
        - FORBIDDEN
        - RESOURCE_NOT_FOUND
        - METHOD_NOT_ALLOWED
        - CONFLICT
        - STAFF_PHONE_ALREADY_EXISTS
        - STAFF_ACCOUNT_UNAVAILABLE
        - STAFF_PHONE_BINDING_REQUIRED
        - STAFF_ACCOUNT_VERSION_CONFLICT
        - WECHAT_IDENTITY_CONFLICT
        - WECHAT_PROVIDER_RESPONSE_INVALID
        - LESSON_NOT_ASSIGNED
        - LESSON_ALREADY_COMPLETED
        - LESSON_VERSION_CONFLICT
        - LESSON_REVERSAL_WINDOW_EXPIRED
        - ATTENDANCE_INCOMPLETE
        - INSUFFICIENT_LESSON_BALANCE
        - FEEDBACK_NOT_ALLOWED
        - FEEDBACK_IMAGE_INVALID
        - FEEDBACK_IMAGE_CONFLICT
        - GROUP_CAMPAIGN_POSTER_INVALID
        - CUSTOMER_SERVICE_QR_INVALID
        - IDEMPOTENCY_CONFLICT
        - EARNING_RULE_NOT_FOUND
        - EARNING_RULE_OVERLAP
        - EARNING_NOT_REVIEWABLE
        - EARNING_ALREADY_REVIEWED
        - EARNING_REVERSED
        - WITHDRAWAL_AMOUNT_INVALID
        - WITHDRAWAL_BALANCE_INSUFFICIENT
        - WITHDRAWAL_FREQUENCY_EXCEEDED
        - WITHDRAWAL_STATUS_CONFLICT
        - WITHDRAWAL_SELF_REVIEW_FORBIDDEN
        - PAYOUT_PROOF_REQUIRED
        - PAYOUT_PROOF_INVALID
        - TOO_MANY_REQUESTS
        - INTERNAL_SERVER_ERROR
...
    LessonSessionStatus:
      type: string
      enum:
        - SCHEDULED
        - IN_PROGRESS
        - COMPLETED
        - REVERSED
        - CANCELLED
    AttendanceStatus:
      type: string
      enum:
        - PRESENT
        - LEAVE
        - ABSENT
...
    TeacherLessonSessionSummary:
      type: object
      required:
        - id
        - version
        - className
        - courseName
        - campusName
        - startsAt
        - endsAt
        - lessonUnits
        - status
        - studentCount
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        className:
          type: string
          minLength: 1
          maxLength: 200
        courseName:
          type: string
          minLength: 1
          maxLength: 200
        campusName:
          type: string
          minLength: 1
          maxLength: 200
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        lessonUnits:
          type: integer
          minimum: 1
          description: Lesson units in hundredths of one lesson.
        status:
          $ref: '#/components/schemas/LessonSessionStatus'
        studentCount:
          type: integer
          minimum: 0
      additionalProperties: false
...
    AttendancePolicyRule:
      type: object
      required:
        - consumesLessonUnits
      properties:
        consumesLessonUnits:
          type: boolean
      additionalProperties: false
    AttendancePolicy:
      type: object
      required:
        - PRESENT
        - LEAVE
        - ABSENT
      properties:
        PRESENT:
          allOf:
            - $ref: '#/components/schemas/AttendancePolicyRule'
            - type: object
              properties:
                consumesLessonUnits:
                  type: boolean
                  const: true
        LEAVE:
          $ref: '#/components/schemas/AttendancePolicyRule'
        ABSENT:
          $ref: '#/components/schemas/AttendancePolicyRule'
      additionalProperties: false
    TeacherLessonStudent:
      type: object
      required:
        - id
        - displayName
        - attendanceStatus
        - expectedConsumeUnits
        - feedback
        - feedbackImages
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        attendanceStatus:
          oneOf:
            - $ref: '#/components/schemas/AttendanceStatus'
            - type: 'null'
        expectedConsumeUnits:
          type: integer
          minimum: 0
        feedback:
          oneOf:
            - type: string
              maxLength: 1000
            - type: 'null'
        feedbackImages:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/FeedbackImageView'
      additionalProperties: false
    TeacherLessonSessionDetail:
      allOf:
        - $ref: '#/components/schemas/TeacherLessonSessionSummary'
        - type: object
          required:
            - attendancePolicy
            - students
            - canComplete
            - canReverse
          properties:
            attendancePolicy:
              $ref: '#/components/schemas/AttendancePolicy'
            students:
              type: array
              items:
                $ref: '#/components/schemas/TeacherLessonStudent'
            canComplete:
              type: boolean
            canReverse:
              type: boolean
...
    AttendanceItem:
      type: object
      required:
        - studentId
        - status
      properties:
        studentId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AttendanceStatus'
      additionalProperties: false
    AttendanceDraft:
      type: object
      required:
        - lessonVersion
        - attendance
      properties:
        lessonVersion:
          type: integer
          minimum: 1
        attendance:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AttendanceItem'
      additionalProperties: false
    CompleteLessonRequest:
      $ref: '#/components/schemas/AttendanceDraft'
    CompleteLessonConsumption:
      type: object
      required:
        - studentId
        - mainUnits
        - giftUnits
      properties:
        studentId:
          type: string
          format: uuid
        mainUnits:
          type: integer
          minimum: 0
        giftUnits:
          type: integer
          minimum: 0
      additionalProperties: false
    CompleteLessonResult:
      type: object
      required:
        - lessonSessionId
        - lessonVersion
        - status
        - teachingRecordId
        - consumed
      properties:
        lessonSessionId:
          type: string
          format: uuid
        lessonVersion:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
            - COMPLETED
            - REVERSED
        teachingRecordId:
          type: string
          format: uuid
        consumed:
          type: array
          items:
            $ref: '#/components/schemas/CompleteLessonConsumption'
      additionalProperties: false
    ReverseLessonRequest:
      type: object
      required:
        - lessonVersion
        - reason
      properties:
        lessonVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
...
    FeedbackImageView:
      type: object
      required:
        - id
        - mimeType
        - sizeBytes
        - accessUrl
        - accessUrlExpiresAt
      properties:
        id:
          type: string
          format: uuid
        mimeType:
          type: string
          enum:
            - image/jpeg
            - image/png
        sizeBytes:
          type: integer
          minimum: 1
          maximum: 2097152
        accessUrl:
          type: string
          minLength: 1
        accessUrlExpiresAt:
          type: string
          format: date-time
      additionalProperties: false
...
    TeacherLessonSessionEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TeacherLessonSessionDetail'
    CompleteLessonEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CompleteLessonResult'
...